Added return values to zts_start() and zts_simple_start(). Also Fixed signed comparison warnings in ztproxy when compiling under Windows

+++ src/VirtualTap.hpp
@@ -206,7 +206,7 @@ namespace ZeroTier {
 		char vtap_abbr_name[16];
 
 		static int devno;
-		int ifindex = 0;
+		size_t ifindex = 0;
 
 		std::vector<InetAddress> ips() const;
 		std::vector<InetAddress> _ips;
